refactor: 인터페이스 기반 컨트롤러를 등록하는 책임을 ManualControllerConfigurator로 분리 후 ManualHandlerMapping을 mvc 패키지 하위로 이동
