+++ app/src/main/java/com/techcourse/DispatcherServletConfigurator.java
@@ -6,13 +6,15 @@ import com.interface21.webmvc.servlet.mvc.tobe.handleradapter.HandlerExecutionAd
 import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.HandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.HandlerMappingRegistry;
+import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.ManualHandlerMapping;
 
 public class DispatcherServletConfigurator {
 
     public static HandlerMappingRegistry handlerMappingRegistry() {
         final HandlerMappingRegistry handlerMappingRegistry = new HandlerMappingRegistry();
 
-        final HandlerMapping manualHandlerMapping = new ManualHandlerMapping();
+        final HandlerMapping manualHandlerMapping = new ManualHandlerMapping(
+                ManualControllerConfigurator.getManualControllers());
         handlerMappingRegistry.register(manualHandlerMapping);
         final HandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping("com.techcourse.controller");
         handlerMappingRegistry.register(annotationHandlerMapping);
@@ -25,7 +27,7 @@ public class DispatcherServletConfigurator {
 
         handlerAdapterRegistry.register(new ControllerHandlerAdapter());
         handlerAdapterRegistry.register(new HandlerExecutionAdapter());
-        
+
         return handlerAdapterRegistry;
     }
 }
